fix(operations): keep the add popup small and the numba norm length-stable
Three review findings on the Bundle A/B integration.

**The "More" fold-out stopped partitioning anything.** Its default groups were
the optional backend ones (SigPy/BART/User/Other), so demoting those packs left
every group in the everyday partition: the popup became a flat 37-row scroll
above an empty fold-out, which is the exact shape the fold-out exists to
prevent. Default everything except the pinned Common section into "More"
instead. That reads aggressive, but Common already holds the head of each group,
so what remains in the groups is the less-used tail -- and anything a user
reaches for often is one checkbox away from Common in the manager, with the
layout able to override the list wholesale. The collapsed popup is nine rows
and a toggle again.

Selecting by id now reveals a folded-away operation rather than returning "not
found" for one that plainly exists; without that, `select_operation` silently
depended on which side of the fold an operation landed on, which would trap a
future search box as well as callers.

**The Numba normalize kernel drifted with the axis length.** It accumulated the
squared magnitudes in float32 and summed sequentially where NumPy sums pairwise,
so the relative difference grew to 2.6e-4 at 1e6 samples -- and because the
kernel only engages once the JIT is warm, the same operation on the same data
returned different values depending on compilation timing. Accumulate in
float64: the loop is memory-bound so the wider add is free (40.8 ms vs NumPy's
238.4 ms on 336x336x272 complex64, unchanged), and the error is now flat in the
axis length and never worse than the fallback it replaces. The test that missed
this used an 8-long axis and a tolerance loose enough to hide it; it now
measures both paths against a float64 ground truth at 4k and 200k. Exact
equality is deliberately *not* asserted -- the kernel is now more accurate than
the float32 NumPy path, so bit-equality could only be restored by putting the
drift back.

**Two manager icons contradicted their actions:** unhide used a browser-reload
glyph that reads as headphones at 18 px, and hiding a system operation showed a
trash can even though hiding is reversible. Use visibility / visibility_off, and
keep the trash only for deleting a user operation, which really does remove
files.

Also folds a redundant popup section test into the collapse/expand round trip.

+++ arrayscope/operations/_numba_native_ops.py
@@ -28,11 +28,19 @@ def _build_kernels():
         rows = data.reshape((-1, data.shape[-1]))
         output = np.empty_like(rows)
         for row in prange(rows.shape[0]):
-            norm_squared = np.float32(0.0)
+            # Accumulate in float64 even though the data is float32/complex64:
+            # NumPy's reference path sums pairwise, this loop sums sequentially,
+            # and a float32 accumulator drifts with the axis length (2.6e-4
+            # relative at 1e6 samples). Since the kernel only engages once the
+            # JIT is warm, that drift would make the same operation on the same
+            # data return different values depending on compilation timing.
+            # The accumulator is not the bottleneck here -- this loop is
+            # memory-bound -- so the wider add is effectively free.
+            norm_squared = np.float64(0.0)
             for column in range(rows.shape[1]):
-                magnitude = abs(rows[row, column])
+                magnitude = np.float64(abs(rows[row, column]))
                 norm_squared += magnitude * magnitude
-            norm = np.sqrt(norm_squared)
+            norm = np.float32(np.sqrt(norm_squared))
             if norm == 0:
                 for column in range(rows.shape[1]):
                     output[row, column] = 0
